% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BG_word_characters.R
\docType{data}
\name{BG_word_characters}
\alias{BG_word_characters}
\title{Words in breath groups.}
\format{
A data frame with 45034 rows and 2 variables:
\describe{
  \item{breath group}{breath group, in words}
  \item{words}{words, in characters}
}
}
\source{
Torre_etal2019
}
\usage{
BG_word_characters
}
\description{
A dataset containing the average word size (in characters) of breath groups.
}
\keyword{datasets}
