% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BG_word_time.R
\docType{data}
\name{BG_word_time}
\alias{BG_word_time}
\title{Words in breath groups.}
\format{
A data frame with 45034 rows and 2 variables:
\describe{
  \item{x}{breath group, in words}
  \item{y}{words, in seconds}
}
}
\source{
Torre_etal2019
}
\usage{
BG_word_time
}
\description{
A dataset containing the average word size (in seconds) of breath groups.
}
\keyword{datasets}
