% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menzerath.R
\name{dmenzerath}
\alias{dmenzerath}
\title{Density function for Menzerath-Altmann law}
\usage{
dmenzerath(x, parameters, method = "MAL")
}
\arguments{
\item{x}{single number or numeric vector}

\item{parameters}{named vector for parameters a, b and c}

\item{method}{one of MAL (default), simplified_1, simplified_2, Milicka_1, Milicka_2, Milicka_4, Milicka_8}
}
\value{
single number or numeric vector
}
\description{
Density function for Menzerath-Altmann law
}
