% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menzerath.R
\name{fit.menzerath}
\alias{fit.menzerath}
\title{Fit a menzerath object}
\usage{
\method{fit}{menzerath}(object, method = "MAL", ...)
}
\arguments{
\item{object}{a menzerath type object}

\item{method}{string Method to perform the fitting, could be one of MAL, simplified_1, simplified_2, Milicka_1, Milicka_2, Milicka_4 or Milicka_8}

\item{...}{Other arguments passed to lm}
}
\description{
Fit a menzerath object
}
