% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menzerath.R
\name{get_parameters}
\alias{get_parameters}
\title{Return parameters of the Menzerath-Altmann law.}
\usage{
get_parameters(x)
}
\arguments{
\item{x}{An object with class lm}
}
\value{
named list with a, b and c
}
\description{
The parameters are estimated after taking logs
}
