% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menzerath.R
\name{menzerath}
\alias{menzerath}
\title{A class to describe and plot data following the Menzerath-Altman law}
\usage{
menzerath(tb = tibble(), x = "x", y = "y")
}
\arguments{
\item{x}{Average size of construct at level Ln}

\item{y}{Size of construct at level Ln-1}
}
\description{
A class to describe and plot data following the Menzerath-Altman law
}
