% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpheme_syllable.R
\docType{data}
\name{morpheme_syllable}
\alias{morpheme_syllable}
\title{Syllable length of Indonesian morphemes.}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{x}{morpheme, in syllable}
  \item{y}{syllable, in phoneme}
}
}
\source{
Altmann_1980
}
\usage{
morpheme_syllable
}
\description{
The means were computed from 13444 morphemes
}
\keyword{datasets}
