% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_syllable_phoneme.R
\docType{data}
\name{word_syllable_phoneme}
\alias{word_syllable_phoneme}
\title{Syllable length in English words.}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{x}{word, in syllable}
  \item{y}{syllable, in phoneme}
}
}
\source{
Altmann_1980
}
\usage{
word_syllable_phoneme
}
\description{
Dataset consisting of 15465010 words originally from Roberts (1965) A statistical linguistic analysis of American English. The Hague, Mouton 1965
}
\keyword{datasets}
