% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_syllable_time.R
\docType{data}
\name{word_syllable_time}
\alias{word_syllable_time}
\title{Syllable length in Bachka-German words.}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{word}{word, in syllable}
  \item{syllable}{syllable, in milliseconds}
}
}
\source{
Altmann_1980
}
\usage{
word_syllable_time
}
\description{
The measurement of syllable duration (in milliseconds) dependent on word length (measured in number of syllables) was performed on 488 words (with 3961 repetitions whose means were used for computation) of spoken texts of the German dialect of Bachka (Yugoslavia).
}
\keyword{datasets}
