% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menzerath.R
\name{menzerath}
\alias{menzerath}
\title{A class to describe and plot data following the Menzerath-Altman law}
\usage{
menzerath(tb = tibble(), x = "x", y = "y")
}
\arguments{
\item{x}{Average size of a construct (L_n) measured in units of its direct constituents}

\item{y}{Size of the constituent (L_{n-1})measured in its direct subconstituents}
}
\description{
A class to describe and plot data following the Menzerath-Altman law
}
